+++ src/main/kotlin/org/codefreak/codefreak/service/SeedDatabaseService.kt
@@ -1,13 +1,14 @@
 package org.codefreak.codefreak.service
 
-import java.io.ByteArrayOutputStream
 import java.time.Instant
 import org.codefreak.codefreak.Env
 import org.codefreak.codefreak.auth.Role
 import org.codefreak.codefreak.entity.Assignment
+import org.codefreak.codefreak.entity.Task
 import org.codefreak.codefreak.repository.AssignmentRepository
 import org.codefreak.codefreak.repository.UserRepository
-import org.codefreak.codefreak.util.TarUtil
+import org.codefreak.codefreak.util.TaskTemplate
+import org.codefreak.codefreak.util.TaskTemplateUtil
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.ApplicationListener
@@ -15,7 +16,6 @@ import org.springframework.context.annotation.Profile
 import org.springframework.context.event.ContextRefreshedEvent
 import org.springframework.core.Ordered
 import org.springframework.core.annotation.Order
-import org.springframework.core.io.ClassPathResource
 import org.springframework.stereotype.Service
 
 /**
@@ -59,43 +59,30 @@ class SeedDatabaseService : ApplicationListener<ContextRefreshedEvent> {
 
     log.info("Seeding database with sample data. Assignments belong to user '${teacher.username}'.")
 
-    val assignment1 = Assignment("C Assignment", teacher, Instant.now().plusSeconds(60), active = true)
-    val assignment2 = Assignment("Java Assignment", teacher, Instant.now(), active = true)
-    assignmentRepository.saveAll(listOf(assignment1, assignment2))
+    val tasks = mutableListOf<Task>()
 
-    ByteArrayOutputStream().use {
-      TarUtil.createTarFromDirectory(ClassPathResource("init/tasks/c-add").file, it)
-      taskTarService.createFromTar(it.toByteArray(), teacher, assignment1)
-    }
-    ByteArrayOutputStream().use {
-      TarUtil.createTarFromDirectory(ClassPathResource("init/tasks/java-add").file, it)
-      taskTarService.createFromTar(it.toByteArray(), teacher, assignment2).also { task ->
-        // set a 1h 1min time limit
-        taskService.saveTask(task)
-      }
-    }
+    TaskTemplate.values().forEach {
+      val templateTar = TaskTemplateUtil.readTemplateTar(it)
+      val task = taskTarService.createFromTar(templateTar, teacher)
 
-    // task pool
-    ByteArrayOutputStream().use {
-      TarUtil.createTarFromDirectory(ClassPathResource("init/tasks/java-add").file, it)
-      taskTarService.createFromTar(it.toByteArray(), teacher)
-    }
-    ByteArrayOutputStream().use {
-      TarUtil.createTarFromDirectory(ClassPathResource("init/tasks/c-add").file, it)
-      taskTarService.createFromTar(it.toByteArray(), teacher)
+      val templateName = it.name.toLowerCase().capitalize()
+      task.title = "Program in $templateName"
+      taskService.saveTask(task)
+      tasks.add(task)
+
+      val assignment = Assignment("$templateName Assignment", teacher, Instant.now(), active = true)
+      assignmentService.saveAssignment(assignment)
+      assignmentService.addTasksToAssignment(assignment, listOf(task))
     }
 
-    ByteArrayOutputStream().use {
-      TarUtil.createTarFromDirectory(ClassPathResource("init/tasks").file, it)
-      assignmentService.createFromTar(it.toByteArray(), teacher) {
-        openFrom = Instant.now()
-        deadline = Instant.now().plusSeconds(60)
-        active = true
-      }.let { result ->
-        if (result.taskErrors.isNotEmpty()) {
-          throw result.taskErrors.values.first()
-        }
-      }
-    }
+    val assignmentWithAllTasks = Assignment(
+      "Sample Assignment",
+      teacher,
+      openFrom = Instant.now(),
+      deadline = Instant.now().plusSeconds(60),
+      active = true
+    )
+    assignmentService.saveAssignment(assignmentWithAllTasks)
+    assignmentService.addTasksToAssignment(assignmentWithAllTasks, tasks)
   }
 }