+++ src/main/kotlin/org/codefreak/codefreak/util/TaskTemplateUtil.kt
@@ -1,0 +1,14 @@
+package org.codefreak.codefreak.util
+
+import org.springframework.core.io.ClassPathResource
+
+enum class TaskTemplate {
+  JAVA, PYTHON, CSHARP, JAVASCRIPT
+}
+
+object TaskTemplateUtil {
+  fun readTemplateTar(template: TaskTemplate): ByteArray {
+    val path = "org/codefreak/templates/${template.name.toLowerCase()}.tar"
+    return ClassPathResource(path).inputStream.use { it.readBytes() }
+  }
+}